Рефакторинг перед мержем
